requirements.txt
# Core dependencies for API Chart Bot Application
requests>=2.31.0
plotly>=5.17.0
pandas>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0

# Optional: for enhanced data processing
numpy>=1.24.0
